Move opening of BP entry sheet to `BloodPressureSummaryView`
