+++ app/src/main/java/org/simple/clinic/summary/bloodpressures/BloodPressureSummaryEvent.kt
@@ -1,0 +1,9 @@
+package org.simple.clinic.summary.bloodpressures
+
+import org.simple.clinic.widgets.UiEvent
+
+sealed class BloodPressureSummaryEvent : UiEvent
+
+object NewBloodPressureClicked : BloodPressureSummaryEvent() {
+  override val analyticsName = "Patient Summary:New BP Clicked"
+}
