+++ app/src/main/java/org/simple/clinic/summary/bloodpressures/BloodPressureSummaryUi.kt
@@ -1,7 +1,9 @@
 package org.simple.clinic.summary.bloodpressures
 
 import org.simple.clinic.bp.BloodPressureMeasurement
+import java.util.UUID
 
 interface BloodPressureSummaryUi {
   fun populateBloodPressures(bloodPressureMeasurements: List<BloodPressureMeasurement>)
+  fun showBloodPressureEntrySheet(patientUuid: UUID)
 }
